# docs/swagger/foedie.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ecommerce Auth API",
    "version": "1.0",
    "contact": {
      "name": "Lovelyo",
      "url": "https://lovelyoyrmia.com",
      "email": "lovelyoyrmia@gmail.com"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "summary": "Get Products",
        "description": "Use this API to get list of products",
        "operationId": "ProductService_GetProducts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbGetProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "ProductService"
        ]
      }
    },
    "/api/v1/products/{pid}": {
      "get": {
        "summary": "Get Product Details",
        "description": "Use this API to get product details",
        "operationId": "ProductService_GetProductDetails",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbProduct"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ProductService"
        ]
      }
    }
  },
  "definitions": {
...
    "pbGetProductResponse": {
      "type": "object",
      "properties": {
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pbProduct"
          }
        }
      }
    },
...
    "pbProduct": {
      "type": "object",
      "properties": {
        "pid": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "stock": {
          "type": "string",
          "format": "int64"
        },
        "price": {
          "type": "string",
          "format": "int64"
        },
        "category": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}